CosChic/backend/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cmake==3.29.3
contourpy==1.2.1
cvzone==1.6.1
cycler==0.12.1
dj-rest-auth==6.0.0
Django==4.0
django-allauth==0.63.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
faiss-cpu==1.8.0
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
jax==0.4.28
jaxlib==0.4.28
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
mysqlclient==2.2.4
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
scipy==1.13.1
six==1.16.0
sounddevice==0.4.7
sqlparse==0.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
